Make thumbnailer support storage backend which has a url not a path
The thumbnail helper used by source does nice things like providing
a default keyword and sets the output format as png when the input
format is png (to make transparency work for example). However not
all backends provide the image object with a path attribute. S3 for
example provides a url attribute which should be used instead. So
we need to modify the code in such a way that it supports both.

We then make it return the url instead of the object itself so we
have to remove the {{ MEDIA_URL }} from the templates where the
thumbnail helper is used.